+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ### v0.3.0 2023-07-XX
 
+- Add converstion function `rgb2color`
 - Add converstion function `RGB2color`
 - Add documentation to functions and package
 - Add script for adding Ipe stypes files to Ipe file